refactor: improve search_sorted_cache, add fast_/sorted_cache files
